dm_bot/requirements.txt
aiogram==3.13.1
pydantic-settings==2.6.1
PyJWT==2.9.0
redis==5.0.7
